data_receiver/requirements.txt
pyserial==3.5
pymongo==4.6.0
